ocr-service-anti/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# OCR
pytesseract>=0.3.10
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# QR Code
pyzbar>=0.1.9

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
nest_asyncio>=1.5.0

# Training (optional)
fonttools>=4.47.0
